requirements.txt
viam-sdk>=0.60.0
adafruit-circuitpython-mlx90640>=1.3.0
adafruit-blinka>=8.60.0
pillow>=10.0.0
